% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epipal_c}
\alias{epipal_c}
\title{Continuous Epinion color palettes}
\format{
A \code{list} with 9 elements
}
\usage{
epipal_c
}
\description{
Continuous Epinion color palettes
}
\keyword{datasets}
