% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epipal_d}
\alias{epipal_d}
\title{Discrete Epinion color palettes}
\format{
A \code{list} with 7 elements
}
\usage{
epipal_d
}
\description{
Discrete Epinion color palettes
}
\keyword{datasets}
