% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitheme_map.R
\name{epitheme_map}
\alias{epitheme_map}
\title{Apply an Epinion styled theme to your \code{ggplot2} graphics}
\usage{
epitheme_map(legend = TRUE, textcolor = "DarkBlue")
}
\arguments{
\item{legend}{a logical value indicating whether a legend is included. Defaults to \code{TRUE} (with \code{legend.position} defaulting to \code{"bottom"})}

\item{textcolor}{Color for text elements in \code{theme()} options. Must be either "black" or a named color from the Epinion color palette (see \link[repinion]{getcols_epi}), with or without the \code{"Epinion"} prefix. Default is "DarkBlue"}
}
\value{
a CVI compliant plot
}
\description{
This function applies a minimalist theme suitable for maps and consistent with the Epinion Corporate Visual Identity (CVI) to your \code{ggplot2} graphics. Add it to your \code{ggplot2} syntax for conistent visual outputs.
}
\examples{
library(tidyverse)
ggplot(mtcars, aes(x = wt, y = mpg)) +
geom_point() +
epitheme_map()
}
\author{
Jeppe Vierø
}
