% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_epi_c.R
\name{fill_epi_c}
\alias{fill_epi_c}
\title{Apply a continuous color scale to your \code{ggplot2} graphics}
\usage{
fill_epi_c(palette = "epiblue", reverse = FALSE, ...)
}
\arguments{
\item{palette}{a continuous color palette. Default is "epiblue"}

\item{reverse}{A logical value indicating whether to reverse the color scale. Defaults to \code{FALSE}}
}
\description{
This function applies a continuous color scale using colors from the Epinion color palette
}
\examples{
library(tidyverse)
ggplot(mtcars, aes(x = wt, y = mpg, fill = disp)) +
geom_point(shape = 21, color = "NA") +
fill_epi_c()
}
\author{
Jeppe Vierø
}
