% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcols_epi.R
\name{getcols_epi}
\alias{getcols_epi}
\title{Extract colors from the Epinion CVI color palette by their name}
\usage{
getcols_epi(...)
}
\arguments{
\item{...}{takes as input the name of the color(s) needed}
}
\value{
the corresponding HEX code
}
\description{
This function extract colors (in HEX codes) from the Epinion color palette by their official name
}
\author{
Jeppe Vierø
}
