% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabcol.R
\name{grabcol}
\alias{grabcol}
\title{Extract colors (HEX) from the Epinion CVI color palette by their name}
\usage{
grabcol(color)
}
\arguments{
\item{color}{an Epinion color (see \code{names(repinion::epi_cols)}) with or without Epinion prefix}
}
\value{
the corresponding HEX code
}
\description{
This function extracts colors (in HEX codes) from the Epinion color palette by their official name
}
\author{
Jeppe Vierø
}
