% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_epi_d.R
\name{color_epi_d}
\alias{color_epi_d}
\title{Apply a discrete color scale to your \code{ggplot2} graphics}
\usage{
color_epi_d(
  palette = "main",
  primary = "DarkBlue",
  secondary = "WarmSand",
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{a continuous color palette. Default is "main"}

\item{primary}{a name of a valid Epinion color (see \link[repinion]{getcols_epi}) to be applied to the primary category. Invalid with more than three levels}

\item{secondary}{a name of a valid Epinion color (see \link[repinion]{getcols_epi}) to be applied to the primary category. Invalid with more than three levels}

\item{reverse}{a logical value indicating whether to reverse the color scale. Defaults to \code{FALSE}}
}
\description{
This function applies a discrete color scale using colors from the Epinion color palette
}
\examples{
library(tidyverse)
ggplot(mtcars, aes(x = wt, y = mpg, color = factor(am))) +
geom_point() +
color_epi_d()
}
\author{
Jeppe Vierø
}
